apps/agentic_app_template/service/pyproject.toml
[project]
name = "agentic-app-template-service"
license = "Apache-2.0"
version = "0.1.0"
authors = [{ name = "Aleph Alpha Engineering", email = "engineering@aleph-alpha.com" }]
description = "Service template"

requires-python = "==3.12.8"

dependencies = [
    "cloudscraper>=1.2.71",
    "jinja2>=3.1.5",
    "js2py>=0.74",
    "litellm>=1.61.20",
    "requests>=2.32.3",
    "smolagents",
    "python-liquid==1.13.0",
    "fastapi==0.115.8",
    "uvicorn==0.34.0",
    "httpx==0.28.1",
    "pydantic==2.10.6",
    "pydantic-settings==2.7.1",
    "prometheus_fastapi_instrumentator==7.0.2",
    "litellm>=1.63.11",
    "smolagents>=1.12.0",
    "colorama>=0.4.6",
    "loguru>=0.7.3",
    "pyjwt==2.10.1",
]

[dependency-groups]
dev = [
    "mypy==1.14.1",
    "pre-commit==4.1.0",
    "pylama[toml]==8.4.1",
    "pytest==8.3.4",
    "types-setuptools==75.8.0.20250110",
    "setuptools==75.8.0",
    "pytest-asyncio==0.25.2",
    "ruff==0.9.4",
    "types-requests>=2.32.0.20250515",
    "types-jwt>=0.1.3",
]

[tool.hatch.build.targets.sdist]
include = ["src/service"]

[tool.hatch.build.targets.wheel]
include = ["src/service"]

[tool.hatch.build.targets.wheel.sources]
"src/service" = "service"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pylama]
skip = ".cache/*,*/__init__.py,.venv/*,.venv/*"
ignore = "E501,E203"

[tool.pylama.linter.pycodestyle]
# https://black.readthedocs.io/en/stable/guides/using_black_with_other_tools.html#pycodestyle
max_line_length = 88
ignore = "E203,E701,E704,W503"

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"

[project.scripts]
dev = "service.main:main"
validate_settings = "service.validate_settings:main"

[tool.mypy]
no_implicit_optional = false
disallow_untyped_defs = true
mypy_path = "src"
explicit_package_bases = true

[[tool.mypy.overrides]]
module = ["colorama", "smolagents.*", "markdownify"]
follow_untyped_imports = true

[tool.ruff.lint]
select = ["I"]
